Rename EnterPaymentInfoPage file
