+++ lib/routes/home/widgets/bottom_actions_bar/send_options_bottom_sheet.dart
@@ -4,7 +4,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:l_breez/cubit/account/account_cubit.dart';
 import 'package:l_breez/routes/chainswap/send/send_chainswap_page.dart';
 import 'package:l_breez/routes/home/widgets/bottom_actions_bar/bottom_action_item_image.dart';
-import 'package:l_breez/routes/home/widgets/bottom_actions_bar/enter_payment_info_dialog.dart';
+import 'package:l_breez/routes/home/widgets/bottom_actions_bar/enter_payment_info_page.dart';
 import 'package:l_breez/theme/theme.dart';
 
 class SendOptionsBottomSheet extends StatefulWidget {
